add, update, delete

+++ src/main/java/com/fariba/springrestdemo/dao/CustomerDAO.java
@@ -10,9 +10,9 @@ public interface CustomerDAO {
 
     void saveCustomer(Customer customer);
 
-    Customer getCustomer(int id);
+    Customer getCustomer(long id);
 
-    void deleteCustomer(int id);
+    void deleteCustomer(long id);
 
     List<Customer> searchCustomers(String searchName);
 }
